rename: 상품 등록 페이지 폴더명 파스칼 케이스로 변경
