+++ src/components/index.js
@@ -44,7 +44,7 @@ export {
 export { Post } from './common/Post';
 export { Comment } from './comment/Comment';
 export { CommentInput } from './comment/CommentInput';
-export { AddProductForm } from './addProduct/addProductForm';
+export { AddProductForm } from './addProduct/AddProductForm';
 export { PostContainer } from './post/PostContainer';
 export { CommentList } from './comment/CommentList';
 export { Follow } from './follow/Follow';
